+++ LyranetworkLyra/src/Sdk/Form/Api.php
@@ -249,7 +249,6 @@ class Api
     {
         return array(
             'AUTHORISED',
-            'AUTHORISED_TO_VALIDATE', // TODO is this a pending status?
             'CAPTURED',
             'ACCEPTED',
             'PARTIALLY_AUTHORISED'
@@ -270,6 +269,8 @@ class Api
             'UNDER_VERIFICATION',
             'PRE_AUTHORISED',
             'WAITING_FOR_PAYMENT',
+            'AUTHORISED_TO_VALIDATE',
+            'SUSPENDED',
             'PENDING'
         );
     }
@@ -280,7 +281,11 @@ class Api
      */
     public static function getCancelledStatuses()
     {
-        return array('ABANDONED');
+        return array(
+            'ABANDONED',
+            'NOT_CREATED',
+            'CANCELLED'
+        );
     }
 
     /**
@@ -289,7 +294,10 @@ class Api
      */
     public static function getToValidateStatuses()
     {
-        return array('WAITING_AUTHORISATION_TO_VALIDATE', 'AUTHORISED_TO_VALIDATE');
+        return array(
+            'WAITING_AUTHORISATION_TO_VALIDATE', 
+            'AUTHORISED_TO_VALIDATE'
+        );
     }
 
     /**