+++ LyranetworkLyra/src/Sdk/Tools.php
@@ -13,6 +13,12 @@ namespace Lyranetwork\Lyra\Sdk;
 
 use Symfony\Component\HttpFoundation\Request;
 
+use Lyranetwork\Lyra\Sdk\Form\Response as LyraResponse;
+use Lyranetwork\Lyra\Sdk\Form\Api as LyraApi;
+
+use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStates;
+use Shopware\Core\System\StateMachine\Aggregation\StateMachineTransition\StateMachineTransitionActions;
+
 class Tools
 {
     private static $GATEWAY_CODE = 'Lyra';
@@ -28,13 +34,22 @@ class Tools
 
     private static $CMS_IDENTIFIER = 'Shopware_6.5.x';
     private static $SUPPORT_EMAIL = 'support-ecommerce@lyra-collect.com';
-    private static $PLUGIN_VERSION = '4.0.0';
+    private static $PLUGIN_VERSION = '4.1.0';
     private static $GATEWAY_VERSION = 'V2';
+    private static $REST_URL = 'https://api.lyra.com/api-payment/';
+    private static $STATIC_URL = 'https://static.lyra.com/static/';
 
     public static $pluginFeatures = [
         'qualif' => false,
         'prodfaq' => false,
-        'shatwo' => true
+        'shatwo' => true,
+        'smartform' => true
+    ];
+
+    public static $smartformModes = [
+        'MODE_SMARTFORM',
+        'MODE_SMARTFORM_EXT_WITH_LOGOS',
+        'MODE_SMARTFORM_EXT_WITHOUT_LOGOS'
     ];
 
     public static function getDefault($name)
@@ -56,5 +71,41 @@ class Tools
         $minor = substr($version, 0, strrpos($version, '.'));
 
         return self::getDefault('GATEWAY_CODE') . '_' . self::getDefault('CMS_IDENTIFIER') . '_v' . $minor . '*.pdf';
+    }
+
+    public static function checkRestIpnValidity($request): bool
+    {
+        return $request->get('kr-hash') !== null && $request->get('kr-hash-algorithm') !== null && $request->get('kr-answer') !== null;
+    }
+
+    public static function getNewOrderPaymentStatus(LyraResponse $lyraResponse, ?string $successStatus)
+    {
+        $status = $lyraResponse->get('trans_status');
+
+        if (in_array($status, LyraApi::getSuccessStatuses(), true)) {
+            return $successStatus;
+        } elseif ($lyraResponse->isPendingPayment()) {
+            return OrderTransactionStates::STATE_OPEN; // Open.
+        } elseif ($lyraResponse->isCancelledPayment()) {
+            return OrderTransactionStates::STATE_CANCELLED; // Cancelled.
+        } else {
+            return OrderTransactionStates::STATE_FAILED; // Failed.
+        }
+    }
+
+    public static function getNewOrderTransition(LyraResponse $lyraResponse)
+    {
+        $status = $lyraResponse->get('trans_status');
+
+        if ($lyraResponse->isPendingPayment()) {
+            return 'pending';
+        } elseif ($lyraResponse->isAcceptedPayment()) {
+            return StateMachineTransitionActions::ACTION_PROCESS;
+        } elseif ($lyraResponse->isCancelledPayment()) {
+            return StateMachineTransitionActions::ACTION_CANCEL;
+        } else {
+            // Case of failed payments.
+            return StateMachineTransitionActions::ACTION_REOPEN;
+        }
     }
 }