+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-# 4.0.0, 2023-12-26:
+# 4.1.0, 2024-01-24:
+- [smartform] Smartform integration.
+- Set return mode to POST by default.
+- Improve order and payment statuses management.
+
+# 4.0.0, 2023-10-30:
 - New plugin for Shopware 6.5.x.